+++ plugin-build/gradle.properties
@@ -1,8 +1,8 @@
-ID=fr.frankois944.spm.kmp.plugin
+ID=fr.frankois944.spmForKmp.plugin
 VERSION=0.0.1
-GROUP=fr.frankois944.spm.kmp
-DISPLAY_NAME=An empty Gradle Plugin from a template
-DESCRIPTION=An empty Gradle plugin created from a template
-WEBSITE=https://github.com/cortinico/kotlin-gradle-plugin-template
-VCS_URL=https://github.com/cortinico/kotlin-gradle-plugin-template
-IMPLEMENTATION_CLASS=fr.frankois944.spm.kmp.plugin.SPMKMPPlugin
+GROUP=fr.frankois944.spmForKmp
+DISPLAY_NAME=Swift Package Manager for Kotlin Multiplaform
+DESCRIPTION=A Gradle plugin for integrating Swift packages into a Kotlin multiplatform module
+WEBSITE=https://github.com/frankois944/spm4Kmp
+VCS_URL=https://github.com/frankois944/spm4Kmp
+IMPLEMENTATION_CLASS=fr.frankois944.spmForKmp.plugin.SPMKMPPlugin
