+++ plugin-build/plugin/src/functionalTest/kotlin/fr/frankois944/spmForKmp/plugin/LocalPackageTest.kt
@@ -1,12 +1,12 @@
-package fr.frankois944.spm.kmp.plugin
+package fr.frankois944.spmForKmp.plugin
 
 import com.autonomousapps.kit.GradleBuilder
 import com.autonomousapps.kit.truth.TestKitTruth.Companion.assertThat
-import fr.frankois944.spm.kmp.plugin.definition.SwiftDependency
-import fr.frankois944.spm.kmp.plugin.fixture.KotlinSource
-import fr.frankois944.spm.kmp.plugin.fixture.SmpKMPTestFixture
-import fr.frankois944.spm.kmp.plugin.fixture.SwiftSource
-import fr.frankois944.spm.kmp.plugin.utils.OpenFolderOnFailureExtension
+import fr.frankois944.spmForKmp.plugin.definition.SwiftDependency
+import fr.frankois944.spmForKmp.plugin.fixture.KotlinSource
+import fr.frankois944.spmForKmp.plugin.fixture.SmpKMPTestFixture
+import fr.frankois944.spmForKmp.plugin.fixture.SwiftSource
+import fr.frankois944.spmForKmp.plugin.utils.OpenFolderOnFailureExtension
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
