+++ plugin-build/plugin/src/main/java/fr/frankois944/spmForKmp/plugin/manifest/TemplateUtils.kt
@@ -1,6 +1,6 @@
-package fr.frankois944.spm.kmp.plugin.manifest
+package fr.frankois944.spmForKmp.plugin.manifest
 
-import fr.frankois944.spm.kmp.plugin.definition.SwiftDependency
+import fr.frankois944.spmForKmp.plugin.definition.SwiftDependency
 
 internal val SwiftDependency.isBinaryDependency: Boolean
     get() =
